+++ src/textual/widgets/_text_area.py
@@ -546,6 +546,19 @@ TextArea {
 
         return highlight_query
 
+    def check_consume_key(self, key: str) -> bool:
+        """Check if the widget may consume the given key.
+
+        As an text area we are expecting to capture printable keys.
+
+        Args:
+            key: A key identifier.
+
+        Returns:
+            `True` if the widget may capture the key in it's `Key` message, or `False` if it won't.
+        """
+        return len(key) == 1 and key.isprintable()
+
     def _build_highlight_map(self) -> None:
         """Query the tree for ranges to highlights, and update the internal highlights mapping."""
         highlights = self._highlights